Update flyweight pattern
